streamlined parallelization.
