+++ src/bldfm/config.py
@@ -1,0 +1,8 @@
+"""
+Provides global variables for BLDFM.
+"""
+
+# Number of threads for parallel computation. 
+# NUM_THREADS=1 equivalent to serial 
+NUM_THREADS = 1
+
